+++ CHANGES.md
@@ -59,6 +59,7 @@ To be released.
      -  (Libplanet.Explorer) Deprecated primary command.
         It will obsoleted in 0.47.0 release.
         You should use `serve` command instead.  [[#2563]]
+ -  (Libplanet.Explorer) Added `schema` subcommand.  [[#2563]]
 
 [#2518]: https://github.com/planetarium/libplanet/issues/2518
 [#2520]: https://github.com/planetarium/libplanet/pull/2520